wall loading...

+++ includes/cub3d.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cub3d.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hbettal <hbettal@student.42.fr>            +#+  +:+       +#+        */
+/*   By: omghazi <omghazi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/10 12:13:42 by omghazi           #+#    #+#             */
-/*   Updated: 2024/09/15 19:14:17 by hbettal          ###   ########.fr       */
+/*   Updated: 2024/09/20 13:15:33 by omghazi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,12 +45,12 @@ void		ft_error(char *str);
 t_parser	*new_parser_node(char *line, int min, int max);
 void    render_2d(t_cub3D *cub);
 void    init_settings(t_cub3D *cub);
-void    bresenhams(int from_x, int from_y, int to_x, int to_y, t_cub3D *cub, uint32_t color);
-void    my_mlx_put_pixel(unsigned int x, unsigned int y, int color, t_cub3D *cub);
+void    bresenhams(long from_x, long from_y, long to_x, long to_y, t_cub3D *cub, uint32_t color);
+void    my_mlx_put_pixel(double x, double y, int color, t_cub3D *cub);
 void    key_handler(mlx_key_data_t key, void* param);
 
 
 /* FUNCTIONS */
-int     wall(t_cub3D *cub, unsigned int x, unsigned int y);
+int     wall(t_cub3D *cub, int x, int y);
 
 #endif
